+++ src/widgets/textedit.cpp
@@ -62,7 +62,7 @@ TextEdit::TextEdit(int index, DWidget *parent)
     QSizeF originSize = QSizeF(m_fontMetric.boundingRect(
                                    "d").width()  + TEXT_MARGIN,  m_fontMetric.boundingRect(
                                    "d").height() + TEXT_MARGIN);
-    this->resize(originSize.width(), originSize.height());
+    this->resize(static_cast<int>(originSize.width()), static_cast<int>(originSize.height()));
     setHorizontalScrollBarPolicy(Qt::ScrollBarAlwaysOff);
     setVerticalScrollBarPolicy(Qt::ScrollBarAlwaysOff);
 
@@ -119,8 +119,8 @@ void TextEdit::updateContentSize(QString content)
 {
     QFontMetricsF fontMetric = QFontMetricsF(this->document()->defaultFont());
     QSizeF docSize =  fontMetric.size(0,  content);
-    this->setMinimumSize(docSize.width() + TEXT_MARGIN, docSize.height() + TEXT_MARGIN);
-    this->resize(docSize.width() + TEXT_MARGIN, docSize.height() + TEXT_MARGIN);
+    this->setMinimumSize(static_cast<int>(docSize.width() + TEXT_MARGIN), static_cast<int>(docSize.height() + TEXT_MARGIN));
+    this->resize(static_cast<int>(docSize.width() + TEXT_MARGIN), static_cast<int>(docSize.height() + TEXT_MARGIN));
     emit  repaintTextRect(this,  QRectF(this->x(), this->y(),
                                         docSize.width() + TEXT_MARGIN, docSize.height() + TEXT_MARGIN));
 }
@@ -202,8 +202,8 @@ void TextEdit::mouseMoveEvent(QMouseEvent *e)
     QPointF movePos = QPointF(posOrigin.x(), posOrigin.y());
 
     if (m_isPressed && movePos != m_pressPoint) {
-        this->move(this->x() + movePos.x() - m_pressPoint.x(),
-                   this->y() + movePos.y() - m_pressPoint.y());
+        this->move(static_cast<int>(this->x() + movePos.x() - m_pressPoint.x()),
+                   static_cast<int>(this->y() + movePos.y() - m_pressPoint.y()));
 
         emit  repaintTextRect(this,  QRectF(qreal(this->x()), qreal(this->y()),
                                             this->width(),  this->height()));
